refactor: move ProtocolMessenger into new package
