+++ dht.go
@@ -20,6 +20,7 @@ import (
 	pb "github.com/libp2p/go-libp2p-kad-dht/pb"
 	"github.com/libp2p/go-libp2p-kad-dht/providers"
 	"github.com/libp2p/go-libp2p-kad-dht/rtrefresh"
+	"github.com/libp2p/go-libp2p-kad-dht/wire"
 	kb "github.com/libp2p/go-libp2p-kbucket"
 	"github.com/libp2p/go-libp2p-kbucket/peerdiversity"
 	record "github.com/libp2p/go-libp2p-record"
@@ -95,7 +96,7 @@ type IpfsDHT struct {
 	ctx  context.Context
 	proc goprocess.Process
 
-	protoMessenger *ProtocolMessenger
+	protoMessenger *wire.ProtocolMessenger
 	messageMgr     *messageManager
 
 	plk sync.Mutex
@@ -193,7 +194,7 @@ func New(ctx context.Context, h host.Host, options ...Option) (*IpfsDHT, error) 
 		strmap:    make(map[peer.ID]*messageSender),
 		protocols: dht.protocols,
 	}
-	dht.protoMessenger, err = NewProtocolMessenger(dht.messageMgr, WithValidator(dht.Validator))
+	dht.protoMessenger, err = wire.NewProtocolMessenger(dht.messageMgr, wire.WithValidator(dht.Validator))
 	if err != nil {
 		return nil, err
 	}
